The script draw different kind of flowers
The script draw different kind of flowers. It prompt user for radius and size of petals plus if petals should over lap or not. Kindly note that I expect user to enter input in integer format plus for Y/N prompt I expect user to enter Just "Y" or "N". I haven't added data validation on inputs as this was not the focus of program

The main focus was to solve the problem where I should be able to draw as many petals as user wish using a generic algorithm. The algorithm should automatically calculate angles, number of petals in flowers etc. I have achieved all these in this script.